Trackster: treat numbers as floats when dynamically filtering. Pack scripts.

+++ static/scripts/trackster.js
@@ -1280,7 +1280,7 @@ extend(NumberFilter.prototype, {
             // No element to filter on.
             return true;
         }
-        var val = parseInt(element[this.index]);
+        var val = parseFloat(element[this.index]);
         return (isNaN(val) || (val >= this.low && val <= this.high));
     },
     /**
